+++ src/components/common/Header.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import Link from 'next/link';
+
+const Header = () => {
+    
+};
+
+export default Header;
